split schema into functional parts
